requirements.txt
aiohttp==3.9.3
pydantic==2.7.1
openai>=1.0.0
python-dotenv
aiogram==3.20.0.post0